massive changes
git-svn-id: svn://localhost/research/software/eclipse_workspace/synthscape@151 3e9724f4-16a3-46c4-b01e-6b039447de01

+++ src/com/synthverse/synthscape/core/AgentFactory.java
@@ -4,6 +4,6 @@ public interface AgentFactory {
 
 	public Agent create();
 
-	public Agent create(Simulation simulation, long generation, long agentId,
+	public Agent create(Simulation simulation, int generation, int agentId,
 			int x, int y);
 }
