+++ src/com/synthverse/synthscape/experiment/test/basic/TestAgentFactory.java
@@ -7,25 +7,25 @@ import com.synthverse.synthscape.core.Simulation;
 
 public class TestAgentFactory implements AgentFactory, Constants {
 
-	private static TestAgentFactory instance = new TestAgentFactory();
+    private static TestAgentFactory instance = new TestAgentFactory();
 
-	private TestAgentFactory() {
-	}
+    private TestAgentFactory() {
+    }
 
-	public static TestAgentFactory getInstance() {
-		return instance;
-	}
+    public static TestAgentFactory getInstance() {
+	return instance;
+    }
 
-	public Agent create() {
-		Agent agent = new TestAgent();
-		return agent;
-	}
+    public Agent create() {
+	Agent agent = new TestAgent();
+	return agent;
+    }
 
-	public Agent create(Simulation simulation, long generation, long agentId,
-			int x, int y) {
-		TestAgent agent = new TestAgent(simulation, SEED_GENERATION_NUMBER,
-				agentId, 200, x, y);
-		return agent;
-	}
+    public Agent create(Simulation simulation, int generation, int agentId,
+	    int x, int y) {
+	TestAgent agent = new TestAgent(simulation, SEED_GENERATION_NUMBER,
+		agentId, 200, x, y);
+	return agent;
+    }
 
 }
